+++ src/event.rs
@@ -1,7 +1,11 @@
 #![cfg(feature = "unstable")]
 
+use nom::bytes::complete::{tag, take_till};
+use nom::character::complete::{char, space0};
+use nom::sequence::delimited;
+
 use crate::pest_parser::Pair;
-use crate::{string, Date};
+use crate::{date::date, string, Date, IResult, Span};
 
 #[derive(Debug, Clone)]
 pub struct Event<'a> {
@@ -33,4 +37,20 @@ impl<'a> Event<'a> {
         let value = string::from_pair(inner.next().expect("no value in event"));
         Self { date, name, value }
     }
+}
+
+pub(crate) fn event(input: Span<'_>) -> IResult<'_, Event<'_>> {
+    let (input, date) = date(input)?;
+    let (input, _) = delimited(space0, tag("event"), space0)(input)?;
+    let (input, name) = delimited(char('"'), take_till(|c| c == '"'), char('"'))(input)?;
+    let (input, _) = space0(input)?;
+    let (input, value) = delimited(char('"'), take_till(|c| c == '"'), char('"'))(input)?;
+    Ok((
+        input,
+        Event {
+            date,
+            name: name.fragment(),
+            value: value.fragment(),
+        },
+    ))
 }