+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 > and is subject to breaking changes.
 
 * Parse `option` directive
+* Parse `event` directive
 
 
 ## [1.15.0] - 2023-03-07